+++ packages/thumbprint-react/components/Image/get-scroll-parent.test.tsx
@@ -14,14 +14,14 @@ test('defaults to `document.body`', () => {
 test('looks for immediate parent with `overflow: scroll`', () => {
     const wrapper = mount(
         <div>
-            <div style={{ overflow: 'scroll' }} data-test="parent">
-                <div data-test="child">goose</div>
+            <div style={{ overflow: 'scroll' }} data-testid="parent">
+                <div data-testid="child">goose</div>
             </div>
         </div>,
     );
 
-    const child = wrapper.find('[data-test="child"]').getDOMNode();
-    const parent = wrapper.find('[data-test="parent"]').getDOMNode();
+    const child = wrapper.find('[data-testid="child"]').getDOMNode();
+    const parent = wrapper.find('[data-testid="parent"]').getDOMNode();
 
     expect(getScrollParent(child)).toBe(parent);
 });
@@ -29,18 +29,18 @@ test('looks for immediate parent with `overflow: scroll`', () => {
 test('looks for grandparent with `overflow: scroll`', () => {
     const wrapper = mount(
         <div>
-            <div style={{ overflow: 'scroll' }} data-test="parent">
+            <div style={{ overflow: 'scroll' }} data-testid="parent">
                 <div>
                     <div>
-                        <p data-test="child">goose</p>
+                        <p data-testid="child">goose</p>
                     </div>
                 </div>
             </div>
         </div>,
     );
 
-    const child = wrapper.find('[data-test="child"]').getDOMNode();
-    const parent = wrapper.find('[data-test="parent"]').getDOMNode();
+    const child = wrapper.find('[data-testid="child"]').getDOMNode();
+    const parent = wrapper.find('[data-testid="parent"]').getDOMNode();
 
     expect(getScrollParent(child)).toBe(parent);
 });
@@ -48,14 +48,14 @@ test('looks for grandparent with `overflow: scroll`', () => {
 test('looks for parents with `overflow: auto`', () => {
     const wrapper = mount(
         <div>
-            <div style={{ overflow: 'auto' }} data-test="parent">
-                <div data-test="child">goose</div>
+            <div style={{ overflow: 'auto' }} data-testid="parent">
+                <div data-testid="child">goose</div>
             </div>
         </div>,
     );
 
-    const child = wrapper.find('[data-test="child"]').getDOMNode();
-    const parent = wrapper.find('[data-test="parent"]').getDOMNode();
+    const child = wrapper.find('[data-testid="child"]').getDOMNode();
+    const parent = wrapper.find('[data-testid="parent"]').getDOMNode();
 
     expect(getScrollParent(child)).toBe(parent);
 });
@@ -63,14 +63,14 @@ test('looks for parents with `overflow: auto`', () => {
 test('looks for parents with `overflow-y: auto`', () => {
     const wrapper = mount(
         <div>
-            <div style={{ overflowY: 'auto' }} data-test="parent">
-                <div data-test="child">goose</div>
+            <div style={{ overflowY: 'auto' }} data-testid="parent">
+                <div data-testid="child">goose</div>
             </div>
         </div>,
     );
 
-    const child = wrapper.find('[data-test="child"]').getDOMNode();
-    const parent = wrapper.find('[data-test="parent"]').getDOMNode();
+    const child = wrapper.find('[data-testid="child"]').getDOMNode();
+    const parent = wrapper.find('[data-testid="parent"]').getDOMNode();
 
     expect(getScrollParent(child)).toBe(parent);
 });
@@ -78,14 +78,14 @@ test('looks for parents with `overflow-y: auto`', () => {
 test('looks for parents with `overflow-x: auto`', () => {
     const wrapper = mount(
         <div>
-            <div style={{ overflowX: 'auto' }} data-test="parent">
-                <div data-test="child">goose</div>
+            <div style={{ overflowX: 'auto' }} data-testid="parent">
+                <div data-testid="child">goose</div>
             </div>
         </div>,
     );
 
-    const child = wrapper.find('[data-test="child"]').getDOMNode();
-    const parent = wrapper.find('[data-test="parent"]').getDOMNode();
+    const child = wrapper.find('[data-testid="child"]').getDOMNode();
+    const parent = wrapper.find('[data-testid="parent"]').getDOMNode();
 
     expect(getScrollParent(child)).toBe(parent);
 });
@@ -95,15 +95,15 @@ test('finds parent and ignores grandparent nodes', () => {
         <div>
             <div style={{ overflow: 'auto' }}>hi</div>
             <div style={{ overflow: 'auto' }}>
-                <div style={{ overflow: 'auto' }} data-test="parent">
-                    <div data-test="child">goose</div>
+                <div style={{ overflow: 'auto' }} data-testid="parent">
+                    <div data-testid="child">goose</div>
                 </div>
             </div>
         </div>,
     );
 
-    const child = wrapper.find('[data-test="child"]').getDOMNode();
-    const parent = wrapper.find('[data-test="parent"]').getDOMNode();
+    const child = wrapper.find('[data-testid="child"]').getDOMNode();
+    const parent = wrapper.find('[data-testid="parent"]').getDOMNode();
 
     expect(getScrollParent(child)).toBe(parent);
 });